+++ src/test/java/one/util/streamex/PairPermutationSpliteratorTest.java
@@ -1,12 +1,12 @@
 /*
  * Copyright 2015, 2016 Tagir Valeev
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *     http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,19 +15,17 @@
  */
 package one.util.streamex;
 
-import java.util.AbstractMap;
+import org.junit.Test;
+
+import java.util.AbstractMap.SimpleEntry;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.Spliterator;
 
-import one.util.streamex.IntStreamEx;
-import one.util.streamex.PairPermutationSpliterator;
-
-import org.junit.Test;
-
-import static one.util.streamex.TestHelpers.*;
-import static org.junit.Assert.*;
+import static one.util.streamex.TestHelpers.checkSpliterator;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 
 /**
  * @author Tagir Valeev
@@ -40,7 +38,7 @@ public class PairPermutationSpliteratorTest {
             int row2 = PairPermutationSpliterator.isqrt(rev);
             assertEquals(row, row2);
         }
-        for (int row : new int[] { 1_000_000_000, 2_000_000_000, Integer.MAX_VALUE - 1, Integer.MAX_VALUE }) {
+        for (int row : new int[]{1_000_000_000, 2_000_000_000, Integer.MAX_VALUE - 1, Integer.MAX_VALUE}) {
             assertEquals(row, PairPermutationSpliterator.isqrt(row * (row + 1L) / 2));
             assertEquals(row - 1, PairPermutationSpliterator.isqrt(row * (row + 1L) / 2 - 1));
         }
@@ -61,10 +59,9 @@ public class PairPermutationSpliteratorTest {
         for (int i : IntStreamEx.rangeClosed(2, 13).boxed()) {
             List<Integer> input = IntStreamEx.range(i).boxed().toList();
             List<Map.Entry<Integer, Integer>> expected = IntStreamEx.range(i)
-                    .<Map.Entry<Integer, Integer>> flatMapToObj(
-                        a -> IntStreamEx.range(a + 1, i).mapToObj(b -> new AbstractMap.SimpleEntry<>(a, b))).toList();
-            checkSpliterator("#" + i, expected, () -> new PairPermutationSpliterator<>(input,
-                    AbstractMap.SimpleEntry<Integer, Integer>::new));
+                    .<Map.Entry<Integer, Integer>>flatMapToObj(
+                            a -> IntStreamEx.range(a + 1, i).mapToObj(b -> new SimpleEntry<>(a, b))).toList();
+            checkSpliterator("#" + i, expected, () -> new PairPermutationSpliterator<>(input, SimpleEntry::new));
         }
     }
 }